setup legit go project structure
