+++ Dockerfile
@@ -12,7 +12,7 @@ WORKDIR /app
  
 # Run command as described:
 # go build will build an executable file named server in the current directory
-RUN go build -o server .
+RUN go build -o rex ./cmd/rex/main.go
  
 # Run the server executable
-CMD [ "/app/server" ]
+CMD [ "/app/rex" ]
