+++ go.mod
@@ -1,8 +1,9 @@
-module rex
+module github.com/root-of-a-tree/remote_exec_bot
 
 go 1.15
 
 require (
 	github.com/bwmarrin/discordgo v0.23.2
-	github.com/labstack/echo/v4 v4.2.0
+	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
+	golang.org/x/sys v0.0.0-20200826173525-f9321e4c35a6 // indirect
 )
